+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Changelog
 - D3 7.0.0 support
 - Remove issue macro, which is now supported by Redmine core
 - new ticket message can be overwritten for projects
+- fixed scope of public project dashboards for all projects
 
 3.0.2
 +++++